# examples/user-management/out/internal/api/handler_organizations.go
// Code generated by stego. DO NOT EDIT.

package api

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"reflect"
	"strconv"
	"strings"

	auth "github.com/example/user-management/out/internal/auth"
	slots "github.com/example/user-management/out/slots"
	"github.com/google/uuid"
)

// OrganizationsHandler handles HTTP requests for Organization entities.
type OrganizationsHandler struct {
	store             Storage
	beforeCreateChain slots.BeforeCreateSlot
}

// NewOrganizationsHandler creates a new OrganizationsHandler.
func NewOrganizationsHandler(store Storage, beforeCreateChain slots.BeforeCreateSlot) *OrganizationsHandler {
	return &OrganizationsHandler{
		store:             store,
		beforeCreateChain: beforeCreateChain,
	}
}

func (h *OrganizationsHandler) Create(w http.ResponseWriter, r *http.Request) {
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	{
		var rawMap map[string]json.RawMessage
		if err := json.Unmarshal(bodyBytes, &rawMap); err == nil {
			if kindRaw, ok := rawMap["kind"]; ok {
				var kind string
				if err := json.Unmarshal(kindRaw, &kind); err != nil || kind != "Organization" {
					handleError(w, r, BadRequest("kind must be Organization"))
					return
				}
			}
		}
	}
	var organization Organization
	if err := json.Unmarshal(bodyBytes, &organization); err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	organization.ID = uuid.New().String()
	if h.beforeCreateChain != nil {
		descriptionVal := ""
		if organization.Description != nil {
			descriptionVal = *organization.Description
		}
		slotReq := &slots.BeforeCreateRequest{
			Input: &slots.CreateRequest{
				Entity: "Organization",
				Fields: map[string]string{
					"name":        organization.Name,
					"description": descriptionVal,
				},
			},
			Caller: func() *slots.Identity {
				id := auth.IdentityFromContext(r.Context())
				return &slots.Identity{UserID: id.UserID, Role: id.Role, Attributes: id.Attributes}
			}(),
		}
		slotResult, slotErr := h.beforeCreateChain.Evaluate(r.Context(), slotReq)
		if slotErr != nil {
			handleError(w, r, InternalError(slotErr.Error()))
			return
		}
		if !slotResult.Ok {
			sc := http.StatusForbidden
			if slotResult.StatusCode > 0 {
				sc = int(slotResult.StatusCode)
			}
			handleError(w, r, errorForStatus(sc, slotResult.ErrorMessage))
			return
		}
		if slotResult.Halt {
			sc := http.StatusOK
			if slotResult.StatusCode > 0 {
				sc = int(slotResult.StatusCode)
			}
			w.WriteHeader(sc)
			return
		}
	}
	if err := h.store.Create(r.Context(), "Organization", organization); err != nil {
		if errors.Is(err, ErrConflict) {
			handleError(w, r, Conflict("resource already exists"))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(presentEntity(organization, "Organization", organization.ID, "/api/user-mgmt/v1/organizations"+"/"+organization.ID))
}

func (h *OrganizationsHandler) Read(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")
	existing, err := h.store.Get(r.Context(), "Organization", id)
	if err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("Organization", id))
		} else {
			handleError(w, r, InternalError(err.Error()))
		}
		return
	}
	existingData, err := json.Marshal(existing)
	if err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	var organization Organization
	if err := json.Unmarshal(existingData, &organization); err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(presentEntity(organization, "Organization", id, "/api/user-mgmt/v1/organizations"+"/"+id))
}

func (h *OrganizationsHandler) Delete(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")
	if err := h.store.Delete(r.Context(), "Organization", id); err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("Organization", id))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func (h *OrganizationsHandler) List(w http.ResponseWriter, r *http.Request) {
	pageStr := r.URL.Query().Get("page")
	pageSizeStr := r.URL.Query().Get("pageSize")
	page, _ := strconv.Atoi(pageStr)
	if page < 1 {
		page = 1
	}
	pageSize, _ := strconv.Atoi(pageSizeStr)
	if pageSize < 1 {
		pageSize = 100
	}
	if pageSize > 65500 {
		pageSize = 65500
	}
	validFields := map[string]bool{"name": true, "description": true}
	orderDir := r.URL.Query().Get("order")
	if orderDir != "" && orderDir != "asc" && orderDir != "desc" {
		handleError(w, r, BadRequest("invalid order value: "+orderDir+"; must be 'asc' or 'desc'"))
		return
	}
	var orderBy []OrderByField
	if orderByStr := r.URL.Query().Get("orderBy"); orderByStr != "" {
		for _, entry := range strings.Split(orderByStr, ",") {
			entry = strings.TrimSpace(entry)
			if entry == "" {
				continue
			}
			parts := strings.Fields(entry)
			fieldName := parts[0]
			dir := "asc"
			if len(parts) > 1 {
				d := strings.ToLower(parts[1])
				if d != "asc" && d != "desc" {
					handleError(w, r, BadRequest("invalid orderBy direction: "+parts[1]+"; must be 'asc' or 'desc'"))
					return
				}
				dir = d
			}
			if !validFields[fieldName] {
				handleError(w, r, BadRequest("invalid orderBy field: "+fieldName))
				return
			}
			orderBy = append(orderBy, OrderByField{Field: fieldName, Direction: dir})
		}
	} else {
		orderBy = []OrderByField{{Field: "created_time", Direction: "desc"}}
	}
	if orderDir != "" {
		for i := range orderBy {
			orderBy[i].Direction = orderDir
		}
	}
	var fields []string
	if fieldsStr := r.URL.Query().Get("fields"); fieldsStr != "" {
		hasID := false
		for _, f := range strings.Split(fieldsStr, ",") {
			f = strings.TrimSpace(f)
			if f == "" {
				continue
			}
			if f != "id" && !validFields[f] {
				handleError(w, r, BadRequest("invalid fields value: "+f))
				return
			}
			fields = append(fields, f)
			if f == "id" {
				hasID = true
			}
		}
		if !hasID {
			fields = append(fields, "id")
		}
	}
	searchExpr := r.URL.Query().Get("search")
	opts := ListOptions{Page: page, Size: pageSize, OrderBy: orderBy, Fields: fields, Search: searchExpr}
	listResult, err := h.store.List(r.Context(), "Organization", "", "", opts)
	if err != nil {
		if errors.Is(err, ErrSearch) {
			handleError(w, r, BadRequest(err.Error()))
		} else {
			handleError(w, r, InternalError(err.Error()))
		}
		return
	}
	w.Header().Set("Content-Type", "application/json")
	hrefBase := "/api/user-mgmt/v1/organizations"
	itemsSlice := reflect.ValueOf(listResult.Items)
	actualSize := itemsSlice.Len()
	presentedItems := make([]map[string]any, actualSize)
	for i := 0; i < actualSize; i++ {
		item := itemsSlice.Index(i).Interface()
		itemID := reflect.ValueOf(item).FieldByName("ID").String()
		itemData, err := json.Marshal(item)
		if err != nil {
			handleError(w, r, InternalError("internal error"))
			return
		}
		var apiItem Organization
		if err := json.Unmarshal(itemData, &apiItem); err != nil {
			handleError(w, r, InternalError("internal error"))
			return
		}
		presentedItems[i] = presentEntity(apiItem, "Organization", itemID, hrefBase+"/"+itemID)
		if len(fields) > 0 {
			allowed := map[string]bool{"id": true, "kind": true, "href": true}
			for _, f := range fields {
				allowed[f] = true
			}
			for k := range presentedItems[i] {
				if !allowed[k] {
					delete(presentedItems[i], k)
				}
			}
		}
	}
	result := map[string]any{
		"kind":  "OrganizationList",
		"page":  page,
		"size":  actualSize,
		"total": listResult.Total,
		"items": presentedItems,
	}
	json.NewEncoder(w).Encode(result)
}